usb: gadget: f_fs: Switch to separate buffers for each bulk endpoint.
This is for trying out with v2 of the proposed patch for FunctionFS mmap.

The main change here is, that the bulk endpoints are now handled completely
separately.
This also means calling mmap to get buffers for each bulk endpoint
separately.

Additionally the maximum perfomance is now a lot higher because the
"bMaxBurst" attribute in the SuperSpeed Endpoint Companion describtor is now
set to the maximum.